+++ src/utils/string.helper.ts
